+++ e2e/commands/link.e2e.js
@@ -5,8 +5,6 @@ import Helper from '../e2e-helper';
 
 chai.use(require('chai-fs'));
 
-const assert = chai.assert;
-
 describe('bit link', function () {
   this.timeout(0);
   const helper = new Helper();
@@ -16,22 +14,50 @@ describe('bit link', function () {
   });
 
   describe('author components', () => {
-    before(() => {
-      helper.reInitLocalScope();
-      const isTypeFixture = "module.exports = function isType() { return 'got is-type'; };";
-      helper.createComponent('utils', 'is-type.js', isTypeFixture);
-      helper.addComponent('utils/is-type.js');
-      const isStringFixture =
-        "const isType = require('bit/utils/is-type/utils/is-type'); module.exports = function isString() { return isType() +  ' and got is-string'; };";
-      helper.createComponent('utils', 'is-string.js', isStringFixture);
-      helper.addComponent('utils/is-string.js');
-      helper.runCmd('bit link');
+    describe('before export', () => {
+      let linkOutput;
+      before(() => {
+        helper.reInitLocalScope();
+        const isTypeFixture = "module.exports = function isType() { return 'got is-type'; };";
+        helper.createComponent('utils', 'is-type.js', isTypeFixture);
+        helper.addComponent('utils/is-type.js');
+        linkOutput = helper.runCmd('bit link');
+      });
+      it('should not create any link because there is no scope yet (until export)', () => {
+        expect(linkOutput).to.have.string('nothing to link because the component was not exported yet');
+        expect(path.join(helper.localScopePath, 'node_modules')).to.not.be.a.path();
+      });
     });
-    it('should print results from the dependency that uses require(bit) syntax', () => {
-      const appJsFixture = "const isString = require('./utils/is-string'); console.log(isString());";
-      fs.outputFileSync(path.join(helper.localScopePath, 'app.js'), appJsFixture);
-      const result = helper.runCmd('node app.js');
-      expect(result.trim()).to.equal('got is-type and got is-string');
+    describe('after export', () => {
+      before(() => {
+        helper.setNewLocalAndRemoteScopes();
+        const isTypeFixture = "module.exports = function isType() { return 'got is-type'; };";
+        helper.createComponent('utils', 'is-type.js', isTypeFixture);
+        helper.addComponent('utils/is-type.js');
+        helper.commitAllComponents();
+        helper.exportAllComponents();
+        const isStringFixture = `const isType = require('@bit/${
+          helper.remoteScope
+        }.utils.is-type/utils/is-type'); module.exports = function isString() { return isType() +  ' and got is-string'; };`;
+        helper.createComponent('utils', 'is-string.js', isStringFixture);
+        helper.addComponent('utils/is-string.js');
+        const appJsFixture = "const isString = require('./utils/is-string'); console.log(isString());";
+        fs.outputFileSync(path.join(helper.localScopePath, 'app.js'), appJsFixture);
+      });
+      it('should generate the links as part of the export process', () => {
+        const result = helper.runCmd('node app.js');
+        expect(result.trim()).to.equal('got is-type and got is-string');
+      });
+      describe('after deleting node_modules and running bit link', () => {
+        before(() => {
+          fs.removeSync(path.join(helper.localScopePath, 'node_modules'));
+          helper.runCmd('bit link');
+        });
+        it('should re-generate the links successfully', () => {
+          const result = helper.runCmd('node app.js');
+          expect(result.trim()).to.equal('got is-type and got is-string');
+        });
+      });
     });
   });
   describe('with custom bind name in bit.json', () => {
@@ -48,13 +74,17 @@ describe('bit link', function () {
     });
     describe('auto linking', () => {
       it('node_modules should contain custom dir name', () => {
-        assert.pathExists(path.join(helper.localScopePath, 'node_modules', 'testLink'));
+        expect(path.join(helper.localScopePath, 'node_modules', 'testLink')).to.be.a.path();
       });
       it('should create symlink inside custom folder', () => {
-        assert.pathExists(path.join(helper.localScopePath, 'node_modules', 'testLink', 'bar', 'foo'));
+        expect(
+          path.join(helper.localScopePath, 'node_modules', 'testLink', `${helper.remoteScope}.bar.foo`)
+        ).to.be.a.path();
       });
       it('should point to generated symlink', () => {
-        expect(path.join(helper.localScopePath, 'node_modules', 'testLink', 'bar', 'foo', 'index.js')).to.be.a.file();
+        expect(
+          path.join(helper.localScopePath, 'node_modules', 'testLink', `${helper.remoteScope}.bar.foo`, 'index.js')
+        ).to.be.a.file();
       });
     });
     describe('manual linking', () => {
@@ -63,13 +93,17 @@ describe('bit link', function () {
         helper.runCmd('bit link');
       });
       it('node_modules should contain custom dir name', () => {
-        assert.pathExists(path.join(helper.localScopePath, 'node_modules', 'testLink'));
+        expect(path.join(helper.localScopePath, 'node_modules', 'testLink')).to.be.a.path();
       });
       it('should create symlink inside custom folder', () => {
-        assert.pathExists(path.join(helper.localScopePath, 'node_modules', 'testLink', 'bar', 'foo'));
+        expect(
+          path.join(helper.localScopePath, 'node_modules', 'testLink', `${helper.remoteScope}.bar.foo`)
+        ).to.be.a.path();
       });
       it('should point to generated symlink', () => {
-        expect(path.join(helper.localScopePath, 'node_modules', 'testLink', 'bar', 'foo', 'index.js')).to.be.a.file();
+        expect(
+          path.join(helper.localScopePath, 'node_modules', 'testLink', `${helper.remoteScope}.bar.foo`, 'index.js')
+        ).to.be.a.file();
       });
     });
   });
@@ -89,9 +123,11 @@ describe('bit link', function () {
       helper.importComponent('bar2/foo2');
     });
     it('node_modules should contain custom dir name', () => {
-      assert.pathExists(path.join(helper.localScopePath, 'node_modules', 'test'));
-      assert.pathExists(path.join(helper.localScopePath, 'node_modules', 'test', 'bar'));
-      assert.pathExists(path.join(helper.localScopePath, 'node_modules', 'test', 'bar2'));
+      expect(path.join(helper.localScopePath, 'node_modules', 'test')).to.be.a.path();
+      expect(path.join(helper.localScopePath, 'node_modules', 'test', `${helper.remoteScope}.bar.foo`)).to.be.a.path();
+      expect(
+        path.join(helper.localScopePath, 'node_modules', 'test', `${helper.remoteScope}.bar2.foo2`)
+      ).to.be.a.path();
     });
   });
   describe('component with dependency tree of 2', () => {
@@ -105,8 +141,10 @@ describe('bit link', function () {
       helper.reInitLocalScope();
       helper.addRemoteScope();
       helper.importComponent('utils/is-type');
-      const isStringFixture =
-        "const isType = require('bit/utils/is-type'); module.exports = function isString() { return isType() +  ' and got is-string'; };";
+      const isStringFixture = `const isType = require('${helper.getRequireBitPath(
+        'utils',
+        'is-type'
+      )}'); module.exports = function isString() { return isType() +  ' and got is-string'; };`;
       helper.createComponent('utils', 'is-string.js', isStringFixture);
       helper.addComponent('utils/is-string.js');
       helper.commitAllComponents();
@@ -117,7 +155,7 @@ describe('bit link', function () {
       helper.importComponent('utils/is-string');
     });
     it('node_modules should contain custom dir name', () => {
-      assert.pathExists(path.join(helper.localScopePath, 'node_modules', 'bitTest'));
+      expect(path.join(helper.localScopePath, 'node_modules', 'bitTest')).to.be.a.path();
     });
   });
   describe('component with dependency tree of 3', () => {
@@ -135,8 +173,9 @@ describe('bit link', function () {
       helper.addRemoteScope();
       helper.modifyFieldInBitJson('bindingPrefix', 'bitTest');
       helper.importComponent('utils/is-type');
-      const isStringFixture =
-        "const isType = require('bitTest/utils/is-type'); module.exports = function isString() { return isType() +  ' and got is-string'; };";
+      const isStringFixture = `const isType = require('bitTest/${
+        helper.remoteScope
+      }.utils.is-type'); module.exports = function isString() { return isType() +  ' and got is-string'; };`;
       helper.createComponent('utils', 'is-string.js', isStringFixture);
       helper.addComponent('utils/is-string.js');
       helper.commitAllComponents();
@@ -147,8 +186,9 @@ describe('bit link', function () {
       helper.addRemoteScope();
       helper.modifyFieldInBitJson('bindingPrefix', 'bitTest2');
       helper.importComponent('utils/is-string');
-      const isStringFixture2 =
-        "const isString = require('bitTest2/utils/is-string'); module.exports = function isString2() { return isString() +  ' and got is-string2'; };";
+      const isStringFixture2 = `const isString = require('bitTest2/${
+        helper.remoteScope
+      }.utils.is-string'); module.exports = function isString2() { return isString() +  ' and got is-string2'; };`;
       helper.createComponent('test', 'is-string2.js', isStringFixture2);
       helper.addComponent('test/is-string2.js');
       helper.commitAllComponents();
@@ -160,24 +200,36 @@ describe('bit link', function () {
       helper.importComponent('test/is-string2');
     });
     it('node_modules should contain custom dir name', () => {
-      assert.pathExists(path.join(helper.localScopePath, 'node_modules', 'bitTest2'));
+      expect(path.join(helper.localScopePath, 'node_modules', 'bitTest2')).to.be.a.path();
     });
     it('node_modules should contain custom dir name2', () => {
-      assert.pathExists(
-        path.join(helper.localScopePath, 'node_modules', 'bitTest2/test/is-string2/node_modules/bitTest2')
-      );
-    });
-    it('node_modules should contain custom dir name3', () => {
-      assert.pathExists(
+      expect(
         path.join(
           helper.localScopePath,
           'node_modules',
-          'bitTest2/test/is-string2/node_modules/bitTest2/utils/is-string/node_modules/bitTest'
+          'bitTest2',
+          `${helper.remoteScope}.test.is-string2`,
+          'node_modules',
+          'bitTest2'
         )
-      );
+      ).to.be.a.path();
+    });
+    it('node_modules should contain custom dir name3', () => {
+      expect(
+        path.join(
+          helper.localScopePath,
+          'node_modules',
+          'bitTest2',
+          `${helper.remoteScope}.test.is-string2/node_modules/bitTest2/${
+            helper.remoteScope
+          }.utils.is-string/node_modules/bitTest`
+        )
+      ).to.be.a.path();
     });
     it('should print results from the dependency that uses require absolute syntax', () => {
-      const appJsFixture = "const isString2 = require('bitTest2/test/is-string2'); console.log(isString2());";
+      const appJsFixture = `const isString2 = require('bitTest2/${
+        helper.remoteScope
+      }.test.is-string2'); console.log(isString2());`;
       fs.outputFileSync(path.join(helper.localScopePath, 'app.js'), appJsFixture);
       const result = helper.runCmd('node app.js');
       expect(result.trim()).to.equal('got is-type and got is-string and got is-string2');