+++ src/consumer/bit-map/bit-map.js
@@ -160,7 +160,7 @@ export default class BitMap {
     allVersions.forEach((version) => {
       if (version !== componentId.toString() && this.components[version].origin !== COMPONENT_ORIGINS.NESTED) {
         logger.debug(`BitMap: deleting an older version ${version} of an existing component ${componentId.toString()}`);
-        delete this.components[version];
+        this._removeFromComponentsArray(version);
       }
     });
   }
@@ -170,7 +170,12 @@ export default class BitMap {
    */
   getExistingComponentId(componentIdStr: string): ?string {
     if (this.components[componentIdStr]) return componentIdStr;
-    if (BitId.parse(componentIdStr).scope) return undefined; // given id has scope, it should have been an exact match
+    const parsedId = BitId.parse(componentIdStr);
+    if (parsedId.scope && !parsedId.hasVersion()) {
+      return Object.keys(this.components).find((component) => {
+        return BitId.parse(component).toStringWithoutVersion() === componentIdStr;
+      });
+    }
     return Object.keys(this.components).find((component) => {
       return BitId.parse(component).toStringWithoutScopeAndVersion() === componentIdStr;
     });
@@ -222,9 +227,7 @@ export default class BitMap {
         );
       }
     } else {
-      // $FlowFixMe
-      this.components[componentIdStr] = { files, origin };
-
+      this.components[componentIdStr] = new ComponentMap({ files, origin });
       this.components[componentIdStr].mainFile = this._getMainFile(
         pathNormalizeToLinux(mainFile),
         this.components[componentIdStr]
@@ -245,10 +248,18 @@ export default class BitMap {
     return this.components[componentIdStr];
   }
 
-  removeComponent(id: string | bitId) {
+  _removeFromComponentsArray(componentId) {
+    const invalidateCache = () => {
+      this.paths = {};
+    };
+    delete this.components[componentId];
+    invalidateCache();
+  }
+
+  removeComponent(id: string | BitId) {
     const bitId = id instanceof BitId ? id : BitId.parse(id);
     const bitmapComponent = this.getExistingComponentId(bitId.toStringWithoutScopeAndVersion());
-    delete this.components[bitmapComponent];
+    this._removeFromComponentsArray(bitmapComponent);
   }
 
   addMainDistFileToComponent(id: string, distFilesPaths: string[]): void {
@@ -299,8 +310,8 @@ export default class BitMap {
     }
     const olderComponentId = olderComponentsIds[0];
     logger.debug(`BitMap: updating an older component ${olderComponentId} with a newer component ${newIdString}`);
-    this.components[newIdString] = R.clone(this.components[olderComponentId]);
-    delete this.components[olderComponentId];
+    this.components[newIdString] = this.components[olderComponentId];
+    this._removeFromComponentsArray(olderComponentId);
   }
 
   getComponent(id: string | BitId, shouldThrow: boolean = false): ComponentMap {