+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- change auto-generated node_modules links to be the same as NPM installation of components (@bit/scope.box.name)
+
 ## [0.12.0-ext.4] - 2017-12-19
 
 - support saving dists files on a pre-configured directory relative to consumer root (rather than saving them inside the component dir)